chore(SOURSD-706): Adds in extra project detail fields, as well as ad…

+++ docker/start.sh
@@ -10,9 +10,11 @@ if [ $APP_ENV = 'local' ] || [ $APP_ENV = 'dev' ]; then
     echo 'running in dev mode - with watch'
     base_command="$base_command --watch"
 
-    # Completely clear down the data in local/dev envs
-    php artisan migrate:fresh
-    php artisan db:seed --class=BaseDemoSeeder
+    if [ $REBUILD_DB = 1 ]; then
+        # Completely clear down the data in local/dev envs
+        php artisan migrate:fresh
+        php artisan db:seed --class=BaseDemoSeeder
+    fi
 else
     # Only forward-facing migrations anywhere else
     php artisan migrate
